packages/QPhaseSDE_cli/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "QPhaseSDE_cli"
version = "0.1.1"
description = "CLI for QPhaseSDE SDE engine"
authors = [ { name = "Your Name" } ]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
	"QPhaseSDE>=0.1.1",
	"typer>=0.9",
	"rich>=13.0",
	"ruamel.yaml>=0.17",
	"pydantic>=2.0",
]

[project.scripts]
qps = "QPhaseSDE_cli.main:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["QPhaseSDE_cli*"]

[project.optional-dependencies]
dev = [
	"pytest>=7.0",
	"ruff>=0.5",
]
